requirements.txt
# Core Machine Learning & AutoML
autogluon==1.4.0

# Web API & Server
fastapi>=0.100.0,<1.0.0
pydantic>=2.0.0,<3.0.0
uvicorn[standard]>=0.22.0

# Data Science & ML Utilities
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0

# Geospatial Utilities
geopy>=2.4.0

# Image Drawing
Pillow>=9.0.0

